Juicebox hic file reader

Read .hic file generated by Juicebox. Currently can be used both local and remote file.

readJuicer(
  file,
  chromosomes = NULL,
  pairs = NULL,
  unit = c("BP", "FRAG"),
  resolution
)

Arguments

file
Filename can be a local path or remote path. The remote path full list can be obtained from http://aidenlab.org/data.html.

chromosomes
A vector contains all the chromosomes. For example c("chr1", "chr2"), the resulting contact matrixs will include all the pairs of interaction (chr1_chr1, chr1_chr2, chr2_chr2).

unit
Unit only supports c("BP", "FRAG"). "BP" means base-pair, and "FRAG" means fragment.

resolution
The desired resolution of the contact matrix. The resolution must be a value from following list.

unit: BP

2500000

1000000

500000

250000

100000

50000

25000

10000

5000

unit: FRAG

500

250

100

50

20

5

2

1

pair
A vector contains all the pair. The pair take format as "1_1" or "chr1_chr1", both means the contact between chromosome1 and chromosome1. If pair presents, chromosomes argument will be ignore.

Value

A list object includes following items.

contact

A list contains all the contact matrix. The keys of list is coded as chrA_chrB.

information

A list contains basic information of the contact matrix.

genomId The genom id of the current hic file.

resolution The list of current hic file available resolution.

chromosomeSizes A dataframe of chromsomeSize (chromsome, size). Can be used in juicer pre function for different genom.

settings

A list contains file name, unit, resolution and chromosomes.

Details

This function is heavily adpoted from both java version juicebox Dump function and c++ version straw.

Examples

library(FreeHiCLite)

## Remote file location. The reomte file include downloading, it may take a while
remoteFilePath = "https://hicfiles.s3.amazonaws.com/hiseq/gm12878/in-situ/combined.hic"

## Local file location
localFilePath = system.file("extdata", "example.hic", package = "FreeHiCLite")

## Chromosomes needs to be extra
chromosomes = c("chr1", "chr2")

## Pairs needs to be extra
pairs = c("1_1", "1_2")
unit = "BP"
resolution = 500000L

## pass chrosomes into function, it will contains all the interaction pairs

dat <- readJuicer(file=localFilePath, chromosomes=chromosomes, pairs = NULL, unit=unit, resolution=resolution)

#> You are trying to access file from: /private/var/folders/bb/9m2pjpgj5nq5xft84vmyv7dm0000gn/T/RtmpZbEQfl/temp_libpath1ad2111c24de/FreeHiCLite/extdata/example.hic. Make sure you provide a valid path.

#> Use pairs: 1_1, 1_2, 2_2

print(names(dat[['contact']]))

#> [1] "1_1" "1_2" "2_2"

## pass pairs into function, it will contains only the given pairs
start = Sys.time()
#dat <- readJuicer(file=localFilePath, chromosomes=chromosomes, pairs = pairs, unit=unit, resolution=resolution) ## Give you same result
dat <- readJuicer(file=localFilePath, chromosomes=NULL, pairs = pairs, unit=unit, resolution=resolution)

#> You are trying to access file from: /private/var/folders/bb/9m2pjpgj5nq5xft84vmyv7dm0000gn/T/RtmpZbEQfl/temp_libpath1ad2111c24de/FreeHiCLite/extdata/example.hic. Make sure you provide a valid path.

#> Use pairs: 1_1, 1_2

end = Sys.time()

print(end - start)

#> Time difference of 0.007586956 secs

str(dat)

#> List of 3
#>  $ contact    :List of 2
#>   ..$ 1_1: int [1:26875, 1:3] 0 0 500000 0 500000 1000000 500000 1000000 1500000 500000 ...
#>   ..$ 1_2: int [1:7551, 1:3] 27000000 34000000 50000000 50500000 57500000 66000000 167500000 175500000 244000000 245000000 ...
#>  $ information:List of 3
#>   ..$ genomId        : chr "hg19"
#>   ..$ resolution     :List of 2
#>   .. ..$ BP  : int [1:9] 2500000 1000000 500000 250000 100000 50000 25000 10000 5000
#>   .. ..$ FRAG: int(0) 
#>   ..$ chromosomeSizes:'data.frame':	26 obs. of  2 variables:
#>   .. ..$ chromosome: chr [1:26] "1" "10" "11" "12" ...
#>   .. ..$ size      : int [1:26] 249250621 135534747 135006516 133851895 115169878 107349540 102531392 90354753 81195210 78077248 ...
#>  $ settings   :List of 4
#>   ..$ unit       : chr "BP"
#>   ..$ resolution : int 500000
#>   ..$ chromosomes: NULL
#>   ..$ file       : chr "/private/var/folders/bb/9m2pjpgj5nq5xft84vmyv7dm0000gn/T/RtmpZbEQfl/temp_libpath1ad2111c24de/FreeHiCLite/extdata/example.hic"
#>  - attr(*, "class")= chr [1:2] "juicer" "freehic"

## Access each contact matrix

contacts = dat[['contact']]

## chromsome 1 vs chromosome 1
key = "1_1"
contactMatrix <- contacts[[key]]
head(contactMatrix)

#>         [,1]    [,2] [,3]
#> [1,]       0       0   19
#> [2,]       0  500000   14
#> [3,]  500000  500000   88
#> [4,]       0 1000000    1
#> [5,]  500000 1000000   29
#> [6,] 1000000 1000000  159

## chromsome 1 vs chromosome 2
key = "1_2"
contactMatrix <- contacts[[key]]
head(contactMatrix)

#>          [,1] [,2] [,3]
#> [1,] 27000000    0    1
#> [2,] 34000000    0    1
#> [3,] 50000000    0    1
#> [4,] 50500000    0    1
#> [5,] 57500000    0    1
#> [6,] 66000000    0    1